Fix bug in Kinetics applyFriction.
Fix a bug in the Kinetics Level Modifier applyFriction method. When
checking to see whether a friction vector entry is larger than the
corresponding entry in the speed vector, the friction value first has to
be multiplied by the delta. Otherwise, a too large or too small value of
what's to be subtracted from the speed vector will be anticipated.

<bug