Rebinded keys for SinglePlayer and Multiplayer
I have added another List to the KeyMapping with actions for SinglePlayer.
In the Singleplayer it is easier to have the action key and movement keys seperated over the range of the keyboard to easily game with two hands (E and arrow keys).
In the Multiplayer you will have different keybindings where hands will be close to each other (P1: Ctrl and arrow keys, P2: V and WAD keys).

<feature