Fix Logger Test testLog method.
The testLog method assumend the log file used for the test will no lines
other than the ones written by the test method. However, when logging is
enabled during the execution of preceding test cases, other lines may be
written to the file used for testLog as well. Therefore, only the last
line of the log should be compared to the one printed by the test case.

<testing,bug